archives/2024/EMG/arduino_recording/requirements.txt
pandas
numpy
openpyxl
pyserial
keyboard
